create_selectivity_sheet Title

Génère une fiche synthétique décrivant les données et contenant un ensemble d’indicateurs de sélectivité.

create_selectivity_sheet(
  data,
  protocol,
  output_dir = NULL,
  output_file = NULL,
  min_length = NULL,
  language = "EN",
  zones = NULL
)

Arguments

data

data.frame A dataframe as generated by prep_sel_data()

protocol

character Type of protocol. “twin” for experiment where both fishing gear are deployed in the same fishing operation (e.g. using twin trawls) ; “paired” for experiment where the two fishing gear are deployed in a set of ‘pairs’ : two different fishing operations, but in very similar conditions; “unpaired” for experiment where the gear are tested independently in two different samples of fishing operations.

output_dir

character Directory name where the sheet should be written

output_file

character Name of the file for the sheet

min_length

numeric Optional. For species analyzed in length, the minimum targeted length for selectivity (e.g., MCRS) for graphics and statistics. A data.frame with a column ‘species’ and a column ‘min_length’ for the corresponding minimum length.

language

character English by default (“EN”). Can be passed in French (“FR”).

zones

character The vector of zones used in the report. Default to NULL uses all the available area

Value

None A Word document is generated in the output_dir as well as all the figures in png. For more details on the selectivity sheet format, see the related vignette : vignette("create-selectivity-sheet", package = "inser")

Examples

# Create tmp folder
output_dir <- tempfile(pattern = "inser")
dir.create(output_dir)

# Setup input OTT data
OTT_data_folder <- system.file("script_origin","Data","Example_OTT", package = "inser")

TR <- readr::read_delim(
  file = file.path(OTT_data_folder, "TR.csv"),
  delim = ";",
  escape_double = FALSE,
  locale = readr::locale(encoding = "WINDOWS-1252"),
  trim_ws = TRUE
)
#> Rows: 2 Columns: 31
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ";"
#> chr (12): sampling_type, landing_country, vessel_flag_country, project, harb...
#> dbl (11): trip_code, year, vessel_length, vessel_power, vessel_size, vessel_...
#> lgl  (8): number_of_sets, days_at_sea, departure_date_time, return_date_time...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.

HH<-read.table(
  file.path(OTT_data_folder, "HH.csv"),
  sep=";",
  header=TRUE,
  encoding = "WINDOWS-1252")#,colClasses = colClasses)

SL<-read.table(
  file.path(OTT_data_folder, "SL.csv"),
  sep=";",
  header=TRUE,
  encoding = "WINDOWS-1252")

HL<-read.table(
  file.path(OTT_data_folder, "HL.csv"),
  sep=";",
  header=TRUE,
  encoding = "WINDOWS-1252")

colClasses<-rep(NA,ncol(HH))
colClasses[which(names(HH)=="statistical_rectangle")]<-"character"

HH<-read.table(
  file.path(OTT_data_folder, "HH.csv"),
  sep=";",
  header=TRUE,
  colClasses = colClasses,
  encoding = "WINDOWS-1252")

# create TAB output
TAB <- prep_sel_data(data=list(TR,HH,SL,HL))

# Setup zone and min_length params
min_length <- data.frame(species=unique(TAB$species),min_length=c(27,24,20,NA,27,NA))
zones <- c("8.a","8.b","7.d","7.e","7.h")

# run function
create_selectivity_sheet(data = TAB,
                         output_dir = output_dir,
                         output_file = "fiche_InseR_twin",
                         protocol = "twin",
                         language = "FR",
                         zones=zones,
                         min_length=min_length)
#> Coordinate system already present. Adding new coordinate system, which will
#> replace the existing one.
#> Coordinate system already present. Adding new coordinate system, which will
#> replace the existing one.
#> Using the FR version of the selectivity template
#> 
#> 
#> processing file: selectivity_sheet_twin.Rmd
#>   |                                                           |                                                   |   0%  |                                                           |.                                                  |   2%                     |                                                           |..                                                 |   4% (unnamed-chunk-1)   |                                                           |...                                                |   6%                     |                                                           |....                                               |   8% (unnamed-chunk-2)   |                                                           |.....                                              |   9%                     |                                                           |......                                             |  11% (unnamed-chunk-3)   |                                                           |.......                                            |  13%                     |                                                           |........                                           |  15% (unnamed-chunk-4)   |                                                           |.........                                          |  17%                     |                                                           |..........                                         |  19% (unnamed-chunk-5)   |                                                           |...........                                        |  21%                     |                                                           |............                                       |  23% (unnamed-chunk-6)   |                                                           |.............                                      |  25% (unnamed-chunk-7)   |                                                           |.............                                      |  26%                     |                                                           |..............                                     |  28% (unnamed-chunk-8)   |                                                           |...............                                    |  30%                     |                                                           |................                                   |  32% (unnamed-chunk-9)   |                                                           |.................                                  |  34%                     |                                                           |..................                                 |  36% (unnamed-chunk-10)  |                                                           |...................                                |  38%                     |                                                           |....................                               |  40% (unnamed-chunk-11)  |                                                           |.....................                              |  42%                     |                                                           |......................                             |  43% (unnamed-chunk-12)  |                                                           |.......................                            |  45%                     |                                                           |........................                           |  47% (unnamed-chunk-13)  |                                                           |.........................                          |  49%                     |                                                           |..........................                         |  51% (unnamed-chunk-14)  |                                                           |...........................                        |  53%                     |                                                           |............................                       |  55% (unnamed-chunk-15)  |                                                           |.............................                      |  57%                     |                                                           |..............................                     |  58% (unnamed-chunk-16)  |                                                           |...............................                    |  60%                     |                                                           |................................                   |  62% (unnamed-chunk-17)  |                                                           |.................................                  |  64%                     |                                                           |..................................                 |  66% (unnamed-chunk-18)  |                                                           |...................................                |  68%                     |                                                           |....................................               |  70% (unnamed-chunk-19)  |                                                           |.....................................              |  72%                     |                                                           |......................................             |  74% (unnamed-chunk-20)  |                                                           |......................................             |  75%                     |                                                           |.......................................            |  77% (unnamed-chunk-21)  |                                                           |........................................           |  79%                     |                                                           |.........................................          |  81% (unnamed-chunk-22)  |                                                           |..........................................         |  83%                     |                                                           |...........................................        |  85% (unnamed-chunk-23)  |                                                           |............................................       |  87%                     |                                                           |.............................................      |  89% (unnamed-chunk-24)  |                                                           |..............................................     |  91%                     |                                                           |...............................................    |  92% (unnamed-chunk-25)  |                                                           |................................................   |  94%                     |                                                           |.................................................  |  96% (unnamed-chunk-26)  |                                                           |.................................................. |  98%                     |                                                           |...................................................| 100% (unnamed-chunk-27)                                                                                                             
#> output file: selectivity_sheet_twin.knit.md
#> /usr/bin/pandoc +RTS -K512m -RTS selectivity_sheet_twin.knit.md --to docx --from markdown+autolink_bare_uris+tex_math_single_backslash --output /tmp/RtmpG6Wmso/inser19b0773d7ba1/fiche_InseR_twin.docx --lua-filter /home/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/pagebreak.lua --highlight-style tango 
#> 
#> Output created: /tmp/RtmpG6Wmso/inser19b0773d7ba1/fiche_InseR_twin.docx

# Clear tmp folder
unlink(output_dir, recursive = TRUE)